ingestion/requirements.txt
requests
kafka-python
pyyaml